razorpay-payment-integration-django/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-phone-field==1.8.1
idna==3.7
python-dotenv==1.0.1
razorpay==1.4.2
requests==2.32.1
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
